+++ src/chart/painteditem.h
@@ -10,15 +10,15 @@ class PaintedItem : public QQuickPaintedItem
 
 protected:
     const struct Padding {
-        qreal   left    = 50,
+        float   left    = 50,
                 right   = 10,
                 top     = 10,
                 bottom  = 20;
     } padding;
 
-    QString format(qreal v);
-    qreal pwidth()  {return width()  - (padding.left + padding.right); }
-    qreal pheight() {return height() - (padding.top  + padding.bottom);}
+    QString format(float v);
+    float pwidth()  {return width()  - (padding.left + padding.right); }
+    float pheight() {return height() - (padding.top  + padding.bottom);}
 
 public:
     PaintedItem (QQuickItem *parent = Q_NULLPTR);
